+++ framework/CHANGELOG.md
@@ -31,7 +31,7 @@ Yii Framework 2 Change Log
 - Enh #12881: Added `removeValue` method to `yii\helpers\BaseArrayHelper` (nilsburg)
 - Enh #12901: Added `getDefaultHelpHeader` method to the `yii\console\controllers\HelpController` class to be able to override default help header in a class heir (diezztsk)
 - Enh: Added constants for specifying `yii\validators\CompareValidator::$type` (cebe)
-
+- Chg #11689: Changed `yii\base\View::beginCache()` and `yii\base\View::renderDynamic()` to work with closures instead of `eval()` (ElisDN)
 
 2.0.10 October 20, 2016
 -----------------------